pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signalflow"
version = "1.0.0"
description = "ASCII call-thread wiring diagram renderer"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Rudolph Pienaar", email = "rudolph.pienaar@childrens.harvard.edu" }
]
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.scripts]
signalflow = "signalflow.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
